Página Location iniciada (apenas a estrutura e estilização basico)
