--- src/components/Home.tsx
@@ -1,17 +1,0 @@
-import Hero from './Hero';
-import BlueSection from './BlueSection';
-import ShowcaseSection from './ShowcaseSection';
-import ServicesSection from './ServicesSection';
-import TeamSection from './TeamSection';
-
-export default function Home() {
-  return (
-    <>
-      <Hero />
-      <BlueSection />
-      <ShowcaseSection />
-      <ServicesSection />
-      <TeamSection />
-    </>
-  );
-}
