--- src/components/Locations.tsx
@@ -1,3 +1,0 @@
-export default function Locations() {
-  return <div style={{padding: '3rem'}}>Locations Page (em construção)</div>;
-}
