+++ src/pages/ContactUs.tsx
@@ -1,6 +1,6 @@
-import ContactHeader from './ContactHeader';
-import ContactIntro from './ContactIntro';
-import ContactForm from './ContactForm';
+import ContactHeader from '../components/ContactHeader';
+import ContactIntro from '../components/ContactIntro';
+import ContactForm from '../components/ContactForm';
 
 export default function ContactUs() {
   return (
