+++ src/pages/Home.tsx
@@ -1,0 +1,17 @@
+import Hero from '../components/Hero';
+import BlueSection from '../components/BlueSection';
+import ShowcaseSection from '../components/ShowcaseSection';
+import ServicesSection from '../components/ServicesSection';
+import TeamSection from '../components/TeamSection';
+
+export default function Home() {
+  return (
+    <>
+      <Hero />
+      <BlueSection />
+      <ShowcaseSection />
+      <ServicesSection />
+      <TeamSection />
+    </>
+  );
+}
