+++ src/pages/Locations.tsx
@@ -1,0 +1,11 @@
+import LocationHeader from "../components/LocationHeader";
+
+export default function Locations() {
+  return (
+      <div style={{ background: '#fff', minHeight: '100vh' }}>
+        <LocationHeader/>
+      </div>
+       )
+} 
+ 
+
